update wallet balance ui
