+++ packages/shared/routes/dashboard/wallet/views/WalletBalance.svelte
@@ -1,11 +1,12 @@
 <script lang="typescript">
-    import { Icon, Text } from 'shared/components'
+    import { Button, Text } from 'shared/components'
+    import { appSettings } from 'shared/lib/appSettings'
     import { walletRoute } from 'shared/lib/router'
     import { WalletRoutes } from 'shared/lib/typings/routes'
     import type { BalanceOverview } from 'shared/lib/wallet'
+    import type { WalletAccount } from 'shared/lib/wallet'
     import { getContext } from 'svelte'
-    import type { Readable } from 'svelte/store'
-    import { appSettings } from 'shared/lib/appSettings'
+    import type { Readable, Writable } from 'svelte/store'
 
     export let locale
     export let color = 'blue' // TODO: profiles will have different colors
@@ -13,11 +14,19 @@
     $: darkModeEnabled = $appSettings.darkMode
 
     const balance = getContext<Readable<BalanceOverview>>('walletBalance')
+    const accounts = getContext<Writable<WalletAccount[]>>('walletAccounts')
+
+    function handleSendClick() {
+        walletRoute.set(WalletRoutes.Send)
+    }
+    function handleReceiveClick() {
+        walletRoute.set(WalletRoutes.Receive)
+    }
 </script>
 
 <style type="text/scss">
     wallet-balance {
-        padding-top: 57px;
+        @apply pt-8;
         padding-bottom: 42px;
         &.compressed {
             padding-top: 32px;
@@ -40,10 +49,12 @@
     class="relative z-0 bg-gradient-to-b from-{color}-500 to-{color}-600 dark:from-gray-800 dark:to-gray-900 rounded-t-xl px-8"
     class:compressed={$walletRoute !== WalletRoutes.Init}>
     <!-- Balance -->
-    <div data-label="total-balance" class="flex flex-col flex-wrap space-y-1.5">
-        <p class="text-11 leading-120 text-white">{locale('general.totalBalance')}</p>
-        <Text type="h2" overrideColor classes="text-white">{$balance.balance}</Text>
-        <Text type="p" overrideColor smaller classes="text-{color}-200">{$balance.balanceFiat}</Text>
+    <div data-label="total-balance" class="flex flex-col flex-wrap space-y-5">
+        <p class="text-11 leading-120 text-white uppercase tracking-widest">{locale('general.balance')}</p>
+        <div class="flex flex-row justify-between items-end">
+            <p class="text-28 leading-120 text-white font-600">{$balance.balance}</p>
+            <p class="text-12 leading-140 text-white font-600">{$balance.balanceFiat}</p>
+        </div>
     </div>
     <img
         class="bg-pattern"
@@ -52,22 +63,12 @@
         src={`assets/patterns/${darkModeEnabled ? 'wallet-balance-darkmode.svg' : 'wallet-balance.svg'}`}
         alt="" />
     {#if $walletRoute === WalletRoutes.Init}
-        <!-- Incoming/Outgoing -->
-        <div data-label="total-movements" class="flex flex-row justify-between mt-8">
-            <div class="flex items-center space-x-4">
-                <Icon boxed icon="small-chevron-down" classes="text-green-600" boxClasses="bg-white dark:bg-gray-900" />
-                <div>
-                    <Text type="p" classes="text-white">{$balance.incoming}</Text>
-                    <Text type="p" overrideColor smaller classes="text-{color}-200">{locale('general.totalIn')}</Text>
-                </div>
+        {#if $accounts.length > 0}
+            <!-- Action Send / Receive -->
+            <div class="flex flex-row justify-between space-x-4 mt-7 mb-3">
+                <Button medium secondary classes="w-full" onClick={handleReceiveClick}>{locale('actions.receive')}</Button>
+                <Button medium secondary classes="w-full" onClick={handleSendClick}>{locale('actions.send')}</Button>
             </div>
-            <div class="flex items-center space-x-4">
-                <Icon boxed icon="small-chevron-up" classes="text-blue-500" boxClasses="bg-white dark:bg-gray-900" />
-                <div>
-                    <Text type="p" classes="text-white">{$balance.outgoing}</Text>
-                    <Text type="p" overrideColor smaller classes="text-{color}-200">{locale('general.totalOut')}</Text>
-                </div>
-            </div>
-        </div>
+        {/if}
     {/if}
 </wallet-balance>
